# open-api-spec.yaml
openapi: 3.0.0
info:
  title: Teacher parent connect API Documentation
  version: 1.0.0
paths:
...
  /students/{id}:
    get:
      summary: Get a student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
    put:
      summary: Update a student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdate'
      responses:
        '200':
          description: The updated student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
    delete:
      summary: Delete a student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The deletion result
          content:
            application/json:
              schema:
                type: object
...
  /teachers/{id}:
    get:
      summary: Get a teacher by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
    put:
      summary: Update a teacher by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherUpdate'
      responses:
        '200':
          description: The updated teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
    delete:
      summary: Delete a teacher by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The deletion result
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        userName:
          type: string
        password:
          type: string
        email:
          type: string
        age:
          type: integer
        addressLine1:
          type: string
        addressCity:
          type: string
        addressState:
          type: string
        pinCode:
          type: integer
        profilePicUrl:
          type: string
        gender:
          type: string
        parent1Name:
          type: string
        parent1Phone:
          type: string
        parent1Email:
          type: string
        parent2Name:
          type: string
        parent2Phone:
          type: string
        parent2Email:
          type: string
        batches:
          type: array
          items:
            type: string
    StudentUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        age:
          type: integer
        addressLine1:
          type: string
        addressCity:
          type: string
        addressState:
          type: string
        pinCode:
          type: integer
        profilePicUrl:
          type: string
        gender:
          type: string
        parent1Name:
          type: string
        parent1Phone:
          type: string
        parent1Email:
          type: string
        parent2Name:
          type: string
        parent2Phone:
          type: string
        parent2Email:
          type: string
        batches:
          type: array
          items:
            type: string
    Teacher:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        userName:
          type: string
        password:
          type: string
        email:
          type: string
        age:
          type: integer
        addressLine1:
          type: string
        addressCity:
          type: string
        addressState:
          type: string
        pinCode:
          type: integer
        profilePicUrl:
          type: string
        gender:
          type: string
        subject:
          type: string
      required:
        - userName
        - email
        - password
    TeacherUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        age:
          type: integer
        addressLine1:
          type: string
        addressCity:
          type: string
        addressState:
          type: string
        pinCode:
          type: integer
        profilePicUrl:
          type: string
        gender:
          type: string
        subject:
          type: string
...